backend/requirement_win.txt
bcrypt
blinker
click
colorama
contourpy
cycler
Flask
Flask-Bcrypt
Flask-Caching
flask-cors
Flask-JWT-Extended
Flask-Login
Flask-Mail
Flask-SQLAlchemy
Flask-WTF
fonttools
greenlet
importlib_metadata
itsdangerous
Jinja2
kiwisolver
MarkupSafe
matplotlib
numpy
packaging
pillow
pygal
PyJWT
pyparsing
python-dateutil
scipy
six
SQLAlchemy
typing_extensions
Werkzeug
WTForms
zipp
Celery
pytz
python-dotenv
redis
billiard
cachelib
click-didyoumean
click-plugins
click-repl
kombu
prompt_toolkit
tzdata
vine
wcwidth
amqp
Brotli
cachetools
certifi
cffi
charset-normalizer
cssselect2
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
httplib2
idna
oauthlib
pandas
proto-plus
protobuf
psutil
pyasn1
pyasn1_modules
pycparser
pydyf
pyphen
PyYAML
requests
requests-oauthlib
rsa
stripe
tinycss2
tinyhtml5
ua-parser
ua-parser-builtins
uritemplate
urllib3
user-agents
weasyprint
webencodings
zopfli

